server/invoice: get seller name and address from settings
