+++ server/polar/kit/email.py
@@ -1,8 +1,11 @@
 import functools
+from typing import Annotated
 
 import email_validator
 from email_validator import EmailNotValidError, caching_resolver
 from email_validator import validate_email as _validate_email
+from pydantic import AfterValidator, EmailStr
+from pydantic_core import PydanticCustomError
 
 from polar.config import settings
 
@@ -15,4 +18,21 @@ validate_email = functools.partial(
     _validate_email, check_deliverability=True, dns_resolver=_email_dns_resolver
 )
 
-__all__ = ["EmailNotValidError", "validate_email"]
+
+def _validate_email_dns(email: str) -> str:
+    try:
+        validate_email(email)
+    except EmailNotValidError as e:
+        raise PydanticCustomError(
+            "value_error",
+            "value is not a valid email address: {reason}",
+            {"reason": str(e)},
+        ) from e
+    else:
+        return email
+
+
+EmailStrDNS = Annotated[EmailStr, AfterValidator(_validate_email_dns)]
+
+
+__all__ = ["EmailNotValidError", "validate_email", "EmailStrDNS"]
